Refactor Delaunay triangulation code
- Move `Point` and `Triangle` structs to a separate module
- Make `Point` fields public
- Make `Triangle::center()` method public
- Make `bowyer_watson()` function public in the new module

This commit refactors the code by moving the `Point` and `Triangle` structs to a separate module called "delaunay". It also makes the fields of the `Point` struct public and exposes the `Triangle::center()` method publicly. Additionally, it makes the `bowyer_watson()` function public in the new module. These changes improve code organization and allow for easier access to relevant functionality.